Long -> Integer

+++ src/main/java/no/einnsyn/apiv3/entities/saksmappe/SaksmappeRepository.java
@@ -5,7 +5,7 @@ import no.einnsyn.apiv3.entities.EinnsynRepository;
 import no.einnsyn.apiv3.entities.enhet.models.Enhet;
 import no.einnsyn.apiv3.entities.saksmappe.models.Saksmappe;
 
-public interface SaksmappeRepository extends EinnsynRepository<Saksmappe, Long> {
+public interface SaksmappeRepository extends EinnsynRepository<Saksmappe, Integer> {
 
   Stream<Saksmappe> findByJournalenhet(Enhet enhet);
 }
