+++ src/main/java/no/einnsyn/apiv3/entities/skjerming/SkjermingRepository.java
@@ -3,5 +3,5 @@ package no.einnsyn.apiv3.entities.skjerming;
 import no.einnsyn.apiv3.entities.EinnsynRepository;
 import no.einnsyn.apiv3.entities.skjerming.models.Skjerming;
 
-public interface SkjermingRepository extends EinnsynRepository<Skjerming, Long> {
+public interface SkjermingRepository extends EinnsynRepository<Skjerming, Integer> {
 }
